requirements.txt
# Dependances directes de l'application
nicegui==3.7.1
pydantic==2.12.5
httpx==0.28.1
python-dotenv==1.2.1

# Outils video (tools/smooth_video.py)
moviepy==2.2.1
